requirements.txt
Flask==3.1.0
gunicorn==21.2.0
pykml==0.2.0
lxml==5.3.2
geojson==3.2.0
python-magic==0.4.27
google-cloud-storage==2.13.0
